Move decoders to the separate file
